Create Hobbies
I created a database of people and hobbies, and each row in hobbies is related to a row in persons via the person_id column.  I  select the 2 tables with a join so that you can see each person's name next to their hobby. Then I added an additional query that shows only the name and hobbies of 'Bobby McBobbyFace', using JOIN combined with WHERE.